+++ dev_appserver_with_datastore.py
@@ -18,6 +18,8 @@ subprocess.call(
             "~/bin/google_appengine/dev_appserver.py"
         ),
         "--host", "0.0.0.0",
+        "--port", "8081",
+        "--admin_port", "8001",
         "--require_indexes", "yes",
         "--storage_path", STORAGEDIR,
         "./app.yaml",